Updated README && Added helpers folder && Did somr refactoring
