+++ README.md
@@ -1,9 +1,9 @@
-# amsterdam-covid-19-numbers-app
+# Amsterdam-covid-19-numbers-app
 
 An application that renders the daily number of people that tested positive on COVID-19 in the municipality of Amsterdam.
 
 Usage:
 
-1. clone repo
-2. npm i
-3. npm start
+1. Download this repository or use `git clone https://github.com/wax-bear/amsterdam-covid-19-numbers-app` in your terminal
+2. Run `npm i`
+3. Run `npm start`
